Add quarantine function
Spins up a container using the given image (default ubuntu:latest)
with GUI and sound capabilities, and the current working dir bind-
mounted inside the container (read-only). A temporary directory is
also created in the current working directory and bind-mounted (w)
inside the container so as to save any necessary data. The direct-
ory is removed if it is empty when the container exits.