Add ~/.config/guix/current/bin to PATH if it exists.
~/.config/guix/current is also a profile much like ~/.guix-config
but only for guix and guix-daemon itself. It is initailized only
after a successful `guix pull`. The reason why there are two prof-
iles: one (~/.guix-config) for normal packages and the other one
(~/.config/guix/current) for guix and guix-daemon itself is to al-
low for the --roll-back of packages without rolling back guix (and
guix-daemon) itself (and vice-versa).