Set LANG to en_US.UTF-8 instead of unset-ing it
This prevents the following issue from manifesting:
https://issues.guix.gnu.org/issue/37662

The logic prevents LANG being set to  en_US.UTF-8  if
it has already been set to anything but than C.UTF-8;
but it seems that $LANG is always C.UTF-8, despite be-
ing set to something else from the shell, while sourc-
ing /etc/profile. So, it seems the trouble of figuring
out the syntax for pure  POSIX  sh  multiple-predicate
conditional was in vain. But still.